bundle scripts: add error checking o scripts that copy pubs and apps

+++ Bundle/fds/scripts/copy_apps.sh
@@ -2,6 +2,7 @@
 app_type=$1
 dir_from=$2
 bot_host=$3
+error_log=$4
 
 if [ "$app_type" != "fds" ]; then
   app_type="smv"
@@ -18,13 +19,24 @@ CP ()
   if [[ "$bot_host" != "" ]] && [[ "$bot_host" != "`hostname`" ]]; then
     scp -q $bot_host:$dir_from/$FROMFILE $dir_to/.
   else
-    cp $dir_from/$FROMFILE $dir_to/.
-  fi
-  if [ -e $dir_to/$FROMFILE ]; then
-    echo $FROMFILE copied to $dir_to
-  else
-    echo ***error: $FROMFILE failed to copy to $dir_to
-    return_code=1
+    if [ -e $dir_from/$FROMFILE ]; then
+      cp $dir_from/$FROMFILE $dir_to/. 
+      if [ -e $dir_to/$FROMFILE ]; then
+        echo $FROMFILE copied to $dir_to
+      else
+        echo "" 
+        echo ***error: $FROMFILE failed to copy to $dir_to
+        echo ""  >> $error_log
+        echo ***error: $FROMFILE failed to copy to $dir_to >> $error_log
+        return_code=1
+      fi
+    else
+      echo ""
+      echo "***error: $dir_from/$FROMFILE does not exist"
+      echo "" >> $error_log
+      echo "***error: $dir_from/$FROMFILE does not exist" $error_log
+      return_code=1
+    fi
   fi
 }
